requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.92.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2~=3.0.3
MarkupSafe~=2.0.1
pydantic==1.10.5
sniffio==1.3.0
SQLAlchemy~=1.4.29
starlette==0.25.0
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp~=3.7.0
passlib~=1.7.4
sqladmin~=0.9.0
PyJWT~=1.7.1
h11~=0.14.0
cryptography~=2.8
pip~=22.1
attrs~=21.4.0
distro~=1.4.0
wheel~=0.34.2
Pillow~=8.4.0
docutils~=0.15.2
tornado~=6.1
PyYAML~=5.3.1
httpcore~=0.16.3
Pygments~=2.10.0
rich~=9.13.0
rfc3986~=1.5.0
certifi~=2021.10.8
setuptools~=60.5.0
requests~=2.22.0
websockets~=10.3
Django~=3.1.2
asgiref~=3.2.10
WTForms~=3.0.1
psutil~=5.9.0
psycopg2~=2.9.1
decorator~=5.1.1
python-multipart~=0.0.6
httpx~=0.23.3
packaging~=21.2
future~=0.18.2
platformdirs~=2.4.1